Remove the masterless state altogether.
That state cannot be used reliably to decided whether the cluster
has a master or not, since it represents the state of the cluster
at the time it was set, not the actual one.

Change the user/DB creation code to cease running when the number
of cluster instances is 0.

Make failover code treat a failover that didn't proceed because
the master node has change as success.