Continue recreating the master pod after the failed failover.
Previously, the failover failure aborted the pod recreation process,
potentially leaving an 'old' version of the former master pod. With
these changes such a failure will not stop the process of recreating
pods.

There is one case when it would be more beneficial to retry the failover,
namely, when the failover failed because of the candidate (that is no
longer healthy, died or not available.) In that case it would be better
to retry with another candidate. However, Patroni API requires matching
text output against a known failure reason, and the chances of candidate's
failure during the pod recreating are not high enough to justify
additional complexity in the operator failure code.