Improvements related to infrastructure roles and logging of naming conflicts.
- when coming across two roles with the same name but different origins simply
  compare the priority of their origins and choose the one with higher
  priority.

Previously, the code employed a special 'hack' for merging infrastructure and
manifest roles with the same name: the result was the infrastructure role, but
with the flags from the manifest one. That hack was necessary because it was
not possible to define flags for the infrastructure roles directly (and,
therefore, it was not possiblw to define, i.e, a superuser infrastructure role
without it). Since it is now possible to define infrastructure role flags
directly via the config map that hack is no longer needed.

- do not accept infrastructure roles without the password

Emit a warning right away. Also, when reading infrastructure roles always try
to read the infrastrure roles configmap, even when we know there are no entries
in the infrastructure roles secret that correspond to those roles defined in
the configmap. That is to emit proper warnings on those roles instead of
ignoring them silently.

- improve debug logging around naming conflicts

Show the origins of conflicting roles and the result.