Address the test failure.
In the first case, the actual test was assumung that the outcome
of merging the infra and the manifest role will be the role with
the flags taken from the manifest one. We no longer merge roles
this way, instead, we overwrite the role with the lower priority
with the one with the higher one (see 2c67633d5). So the test has
been changed as well.

In the latter case, the test was checking what would happen with
two API roles with the same name. The original code preferered the
latter role, the changed code preferred the original one. Since it
is not really a realistic use case (the teams API is supposed to
return unique roles) it doesn't really matter, but for the sake
of consistency I changed the priority check to always prefer the
latter role when comparing those with the same origins.