Use generated code to set status of CRD objects as well.
Right now we set the status with Update, however, Kubernetes 1.11
introduces the /status subresources, allowing us to set the status with
the special updateStatus call in the future. For now, we keep the code
that is compatible with earlier versions of Kubernetes.

As an upside, remove a lot of custom code for building the REST client
for CRDs, as this is now handled by the autogenerated API. We could
remove the custom unmarshaller as well, once we figure out how CRD
validation.

Also renamed postgresql.go to the database go (code that works with
databases, to avoid confusion with the code that deals with PostgreSQL
objects) and status.go to logs_and_api.go to distinguish between the
code that deals with PostgreSQL object status and the one that performs
logging and API utilities.