Add binries_path_template option.
The operator expects that PostgreSQL binaries are located at
/usr/lib/postgresql/<pgVersion>/bin. While this is true for Debian and
Ubuntu based images, it doesn't work for images that uses another binary
location. One example is CentOS where binaries are located at
/usr/pgsql-<pgVersion>/bin.

This commit adds a configuration option to make the bin_path template
parameter configurable. If "binaries_path_location" is set it will be
used to generate the bin_path option. If binaries_path_location is
omitted it will still use the build-in default
/usr/lib/postgresql/<pgVersion>/bin.