Code utilization of syncConnectionPooler
create and sync connectionPooler were doing essentially the same task
with respect to creating new connectionPooler if not present. Hence, the
code from create is already present in sync, which is now utilized and
duplicate code is removed.

Other code duplication is also removed in sync, for deciding upon when
to delete the connectionPooler. Basically, anytime
newNeedConnectionPooler is false, we have to delete it, if present.

Other respective modifications in tests.