Update username.md
I had the opportunity to analyze the case with cluster recovery from the S3 bucket, and until now I have received many questions about this case. Users do not consider the documentation to be explicit enough for such a case and face difficulties. Therefore, I propose to add this description to the documentation.